compose_containers/dispatcher_engine/requirements.txt
bcrypt==4.0.1
email-validator==1.3.1
fastapi==0.94.1
passlib==1.7.4
Pillow==10.0.0
pydub==0.25.1
PyJWT==2.6.0
pypdfium2==4.18.0
python-multipart==0.0.6
requests==2.31.0
uvicorn==0.21.0
pymongo==4.5.0
httpx==0.27.2
